fetch_backend/requirements.txt
django
djangorestframework
drf-yasg
psycopg2-binary